polish(breaking): Rename AnsiStyle to TerminalStyle
"Ansi" is a bit technical, and not less immediately clear what the intended
purpose of the function is. "Terminal" makes it clear this is related
to terminals in some way.

It also is conveivable that there are non-ANSI common or standardized ways
to style character terminal output, so making that implementation detail
part of the name may end up being misleading.